Add special Victory conditions for campaign scenarios
close #3093
